+++ clients/src/main/java/org/apache/kafka/clients/consumer/RoundRobinAssignor.java
@@ -18,8 +18,8 @@ package org.apache.kafka.clients.consumer;
 
 import org.apache.kafka.clients.consumer.internals.AbstractPartitionAssignor;
 import org.apache.kafka.common.TopicPartition;
-import org.apache.kafka.common.utils.CircularIterator;
 import org.apache.kafka.common.utils.Utils;
+import org.apache.kafka.common.utils.internals.CircularIterator;
 
 import java.util.ArrayList;
 import java.util.HashMap;
