+++ clients/src/main/java/org/apache/kafka/common/record/internal/FileRecords.java
@@ -19,8 +19,8 @@ package org.apache.kafka.common.record.internal;
 import org.apache.kafka.common.KafkaException;
 import org.apache.kafka.common.network.TransferableChannel;
 import org.apache.kafka.common.record.internal.FileLogInputStream.FileChannelRecordBatch;
-import org.apache.kafka.common.utils.AbstractIterator;
 import org.apache.kafka.common.utils.Utils;
+import org.apache.kafka.common.utils.internals.AbstractIterator;
 
 import java.io.Closeable;
 import java.io.File;
