+++ clients/src/main/java/org/apache/kafka/common/record/internal/RecordBatch.java
@@ -18,7 +18,7 @@ package org.apache.kafka.common.record.internal;
 
 import org.apache.kafka.common.record.TimestampType;
 import org.apache.kafka.common.utils.BufferSupplier;
-import org.apache.kafka.common.utils.CloseableIterator;
+import org.apache.kafka.common.utils.internals.CloseableIterator;
 
 import java.nio.ByteBuffer;
 import java.util.Iterator;
