+++ clients/src/main/java/org/apache/kafka/common/record/internal/Records.java
@@ -16,7 +16,7 @@
  */
 package org.apache.kafka.common.record.internal;
 
-import org.apache.kafka.common.utils.AbstractIterator;
+import org.apache.kafka.common.utils.internals.AbstractIterator;
 
 import java.util.Iterator;
 import java.util.Optional;
