+++ clients/src/main/java/org/apache/kafka/common/utils/internals/CircularIterator.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.kafka.common.utils;
+package org.apache.kafka.common.utils.internals;
 
 import java.util.Collection;
 import java.util.ConcurrentModificationException;
