+++ clients/src/test/java/org/apache/kafka/common/record/internal/MemoryRecordsTest.java
@@ -25,8 +25,8 @@ import org.apache.kafka.common.record.TimestampType;
 import org.apache.kafka.common.record.internal.MemoryRecords.RecordFilter;
 import org.apache.kafka.common.record.internal.MemoryRecords.RecordFilter.BatchRetention;
 import org.apache.kafka.common.utils.BufferSupplier;
-import org.apache.kafka.common.utils.CloseableIterator;
 import org.apache.kafka.common.utils.Utils;
+import org.apache.kafka.common.utils.internals.CloseableIterator;
 import org.apache.kafka.test.TestUtils;
 
 import org.junit.jupiter.api.Test;
