+++ connect/api/src/main/java/org/apache/kafka/connect/header/ConnectHeaders.java
@@ -16,7 +16,7 @@
  */
 package org.apache.kafka.connect.header;
 
-import org.apache.kafka.common.utils.AbstractIterator;
+import org.apache.kafka.common.utils.internals.AbstractIterator;
 import org.apache.kafka.connect.data.Date;
 import org.apache.kafka.connect.data.Decimal;
 import org.apache.kafka.connect.data.Schema;
