+++ gradle/spotbugs-exclude.xml
@@ -430,7 +430,7 @@ For a detailed description of spotbugs bug categories, see https://spotbugs.read
 
     <Match>
         <!-- False positive - the field is used in other modules and its unit test -->
-        <Class name="org.apache.kafka.common.utils.PrimitiveRef$LongRef"/>
+        <Class name="org.apache.kafka.common.utils.internals.PrimitiveRef$LongRef"/>
         <Field name="value"/>
         <Bug pattern="URF_UNREAD_PUBLIC_OR_PROTECTED_FIELD"/>
     </Match>
