+++ storage/src/main/java/org/apache/kafka/storage/internals/log/Cleaner.java
@@ -27,9 +27,9 @@ import org.apache.kafka.common.record.internal.MutableRecordBatch;
 import org.apache.kafka.common.record.internal.Record;
 import org.apache.kafka.common.record.internal.RecordBatch;
 import org.apache.kafka.common.utils.BufferSupplier;
-import org.apache.kafka.common.utils.CloseableIterator;
 import org.apache.kafka.common.utils.LogContext;
 import org.apache.kafka.common.utils.Time;
+import org.apache.kafka.common.utils.internals.CloseableIterator;
 import org.apache.kafka.storage.internals.utils.Throttler;
 
 import org.slf4j.Logger;
